Feat/selector computed behaviours (#49)
Makes selectors have to opt in to reactions from ancestors. Example, if deleting or replacing an object that also includes the entity we were selecting for, from now on the selector will not run. It will only run when the key listened to really changes. 

We also allow the option for selectors to react to changes from ancestor operations via a optional flag `reactToAncestorChanges`